Rewrite mapping table to be more maintainable. Also includes performance fixes
